+++ CHANGELOG.md
@@ -460,3 +460,4 @@
 * Split up CoreParams into two files; things that aren't game-specific settings are now in CoreParamsBase
 * Updated Unity to 2020.3.36f1
 * Updated Visual Studio Editor package to 2.0.16
+* Implemented backtick option for string sub literal strings